requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==10.3.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.45.0
transformers==4.39.3
